address comments

+++ functions.sql
@@ -78,7 +78,7 @@ $$;
    Adapted from https://github.com/mapbox/postgis-vt-util/blob/master/src/Z.sql
    Intended usage:
      WHERE Z(!scale_denominator!) < 17 */
-CREATE OR REPLACE FUNCTION Z(scale_denominator double precision)
+CREATE OR REPLACE FUNCTION Z(scale_denominator numeric)
   RETURNS integer
   LANGUAGE SQL
   IMMUTABLE PARALLEL SAFE
@@ -87,6 +87,6 @@ AS $$
 SELECT
 	CASE
 		WHEN scale_denominator <= 0 OR scale_denominator > 600000000 THEN NULL
-		ELSE CAST(ROUND(LOG(2, (559082264.028 / scale_denominator)::numeric)) AS integer)
+		ELSE CAST(ROUND(LOG(2, 559082264.028 / scale_denominator)) AS integer)
 	END
 $$;
